+++ sos-webservices/sos-webservices-json/src/main/java/com/sos/joc/model/xmleditor/read/other/AnswerConfiguration.java
@@ -32,8 +32,14 @@ import org.apache.commons.lang.builder.ToStringBuilder;
 })
 public class AnswerConfiguration {
 
+    /**
+     * non negative long
+     * <p>
+     * 
+     * 
+     */
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
     /**
      * 
      * (Required)
@@ -70,13 +76,25 @@ public class AnswerConfiguration {
     @JsonPropertyDescription("Value is UTC timestamp in ISO 8601 YYYY-MM-DDThh:mm:ss.sZ or empty")
     private Date modified;
 
+    /**
+     * non negative long
+     * <p>
+     * 
+     * 
+     */
     @JsonProperty("id")
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
+    /**
+     * non negative long
+     * <p>
+     * 
+     * 
+     */
     @JsonProperty("id")
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
